updating helmholtz drivers to be forward compatible
