+++ applications/helmholtz/nn_training/helmholtz_training.py
@@ -31,6 +31,8 @@ from hessianlearn import *
 
 from neuralNetworks import *
 
+from helmholtz_utilities import *
+
 tf.set_random_seed(1)
 
 # Parse run specifications
@@ -80,164 +82,7 @@ if args.train_data_size <= 256:
 	settings['hess_batch_size'] = int(args.batch_size/32)
 
 ################################################################################
-# Instantiate data
 
-def load_helmholtz_data(data_dir,rescale = False,n_data = np.inf):
-	assert os.path.isdir(data_dir)
-	data_files = os.listdir(data_dir)
-	data_files = [data_dir + file for file in data_files]
-
-	m_files = []
-	q_files = []
-	for file in data_files:
-		if 'ms_on_rank_' in file:
-			m_files.append(file)
-		if 'qs_on_rank_' in file:
-			q_files.append(file)
-
-	ranks = [int(file.split(data_dir+'ms_on_rank_')[-1].split('.npy')[0]) for file in m_files]
-	max_rank = max(ranks)
-
-	# Serially concatenate data
-	m_data = np.load(data_dir+'ms_on_rank_0.npy')
-	q_data = np.load(data_dir+'qs_on_rank_0.npy')
-	for i in range(1,max_rank+1):
-			appendage_m = np.load(data_dir+'ms_on_rank_'+str(i)+'.npy')
-			m_data = np.concatenate((m_data,appendage_m))
-			appendage_q = np.load(data_dir+'qs_on_rank_'+str(i)+'.npy')
-			q_data = np.concatenate((q_data,appendage_q))
-
-	print('m_data.shape = ',m_data.shape)
-	print('q_data.shape = ',q_data.shape)
-
-	if n_data < np.inf:
-		assert type(n_data) is int
-		m_data = m_data[:n_data]
-		q_data = q_data[:n_data]
-	if rescale:
-		from sklearn import preprocessing
-		m_data = preprocessing.scale(m_data)
-		q_data = preprocessing.scale(q_data)
-
-
-	print('m_data.shape = ',m_data.shape)
-	print('q_data.shape = ',q_data.shape)
-	return [m_data,q_data]
-
-
-def get_projectors(data_dir,as_input_tolerance=1e-4,as_output_tolerance=1e-4,\
-					kle_tolerance = 1e-4,pod_tolerance = 1e-4,\
-					 fixed_input_rank = 0, fixed_output_rank = 0, mixed_output = True, verbose = True):
-	projector_dictionary = {}
-	################################################################################
-	# Derivative Informed Input Subspace
-	AS_input_projector = np.load(data_dir+'AS_input_projector.npy')
-	if verbose:
-		print('AS input projector shape before truncation = ', AS_input_projector.shape)
-	if fixed_input_rank > 0:
-		AS_input_projector = AS_input_projector[:,:fixed_input_rank]
-	else:
-		AS_input_d = np.load(data_dir+'AS_d_GN.npy')
-		AS_input_projector = AS_input_projector[:,np.where(AS_input_d>as_input_tolerance)[0]]
-	if verbose:
-		print('AS input projector shape after truncation = ', AS_input_projector.shape)
-	projector_dictionary['AS_input'] = AS_input_projector
-	################################################################################
-	# Derivative Informed Output Subspace
-	AS_output_projector = np.load(data_dir+'AS_output_projector.npy')
-	
-	if verbose:
-		print('AS output projector shape before truncation = ', AS_output_projector.shape)
-	if fixed_output_rank > 0:
-		AS_output_projector = AS_output_projector[:,:fixed_output_rank]
-	else:
-		AS_output_d = np.load(data_dir+'AS_d_NG.npy')
-		AS_output_projector = AS_output_projector[:,np.where(AS_output_d>as_output_tolerance)[0]]
-	if verbose:
-		print('AS output projector shape after truncation = ', AS_output_projector.shape)
-	projector_dictionary['AS_output'] = AS_output_projector
-	################################################################################
-	# KLE Input Subspace
-	KLE_projector = np.load(data_dir+'KLE_projector.npy')
-	if verbose:
-		print('KLE projector shape before truncation = ', KLE_projector.shape)
-	if fixed_input_rank > 0:
-		KLE_projector = KLE_projector[:,:fixed_input_rank]
-	else:
-		KLE_d = np.load(data_dir+'KLE_d.npy')
-		KLE_projector = KLE_projector[:,np.where(KLE_d>kle_tolerance)[0]]
-	if verbose:
-		print('KLE projector shape after truncation = ', KLE_projector.shape)
-	projector_dictionary['KLE'] = KLE_projector
-	################################################################################
-	# POD Output Subspace
-	POD_projector = np.load(data_dir+'POD_projector.npy')
-	if verbose:
-		print('POD projector shape before truncation = ', POD_projector.shape)
-	if fixed_output_rank > 0:
-		POD_projector = POD_projector[:,:fixed_output_rank]
-	else:
-		POD_d = np.load(data_dir+'POD_d.npy')
-		POD_projector = POD_projector[:,np.where(POD_d>pod_tolerance)[0]]
-	if verbose:
-		print('POD projector shape after truncation = ', POD_projector.shape)
-	projector_dictionary['POD'] = POD_projector
-	return projector_dictionary
-
-def modify_projectors(projectors,input_subspace,output_subspace):
-	# Modify the input projectors
-	assert input_subspace in ['kle','as','random']
-
-	if input_subspace in ['kle','as']:
-		# Always orthogonalize AS and KLE for best results
-		orthogonalize_input = True
-		rescale_input = True
-		if input_subspace == 'kle':
-			input_projector = projectors['KLE']
-		elif input_subspace == 'as':
-			input_projector = projectors['AS_input']
-
-
-		if orthogonalize_input:
-			input_projector,_ = np.linalg.qr(input_projector)
-
-		if rescale_input:
-			# Scaling factor of 10 seemed to perform well for KLE and AS
-			# and this was independent of the projector rank.
-			scale_factor_input = 0.05*float(input_projector.shape[0])/(32*float(input_projector.shape[-1]))
-			input_projector /= scale_factor_input*np.linalg.norm(input_projector)
-
-	elif input_subspace == 'random':
-		input_projector = np.random.randn(*projectors['KLE'].shape)
-		input_projector,_ = np.linalg.qr(input_projector)
-		scale_factor_input = 0.05*float(input_projector.shape[0])/(32*float(input_projector.shape[-1]))
-		input_projector /= scale_factor_input*np.linalg.norm(input_projector)
-
-	# Modify the output projectors
-	# It seems that (re)-orthogonalizing the POD vectors
-	# may not improve the neural network.
-	assert output_subspace in ['pod','as','random']
-
-	if output_subspace in ['pod','as']:
-		orthogonalize_output = True
-		rescale_output = True
-		if output_subspace == 'pod':
-			output_projector = projectors['POD']
-		elif output_subspace == 'as':
-			output_projector = projectors['AS_output']
-
-		if orthogonalize_output:
-			output_projector,_ = np.linalg.qr(output_projector)
-
-		if rescale_output:
-			scale_factor_output = 1.
-			output_projector /= scale_factor_output*np.linalg.norm(output_projector)
-
-	if output_subspace == 'random':
-		output_projector = np.random.randn(*projectors['POD'].shape)
-		output_projector /= np.linalg.norm(output_projector)
-
-	return input_projector, output_projector
 
 ntargets = 100
 gamma = args.gamma
@@ -253,13 +98,13 @@ data_dir = '../data/'+formulation+'_n_obs_'+str(ntargets)+'_g'+str(gamma)+'_d'+s
 
 assert os.path.isdir(data_dir), 'Directory does not exist'+data_dir
 
-m_data, q_data = load_helmholtz_data(data_dir,rescale = False,n_data = n_data)
+helmholtz_data = load_helmholtz_data(data_dir,rescale = False,n_data = n_data)
+
+m_data = helmholtz_data['m_data']
+q_data = helmholtz_data['q_data']
 
 input_dim = m_data.shape[-1]
 output_dim = q_data.shape[-1]
-
-# Instante the data object
-data = Data([m_data,q_data],settings['batch_size'],test_data_size = settings['test_data_size'],hessian_batch_size = settings['hess_batch_size'])
 
 
 ################################################################################
@@ -365,10 +210,14 @@ else:
 regressor.summary()
 
 ################################################################################
-# Instantiate the problem, regularization.
+# Instantiate the problem, data and regularization.
 q_mean = np.mean(q_data,axis = 0)
 
 problem = RegressionProblem(regressor,y_mean = q_mean,dtype=tf.float32)
+
+# Instante the data object
+data = Data({problem.x:m_data,problem.y_true:q_data},settings['batch_size'],\
+	validation_data_size = settings['test_data_size'],hessian_batch_size = settings['hess_batch_size'])
 
 settings['tikhonov_gamma'] = 0.0
 
@@ -394,9 +243,9 @@ if 'projected' in architecture and set_weights:
 	HLModelSettings['layer_weights'] = {'input_proj_layer':[input_projector],\
 					'output_layer':[output_projector.T,np.zeros(output_projector.T.shape[-1])]}
 
-HLModelSettings['printing_items'] = {'sweeps':'sweeps','Loss':'loss_train','acc train':'accuracy_train',\
-												'||g||':'||g||','Loss test':'loss_test','acc test':'accuracy_test',\
-												'maxacc test':'max_accuracy_test','var red':'variance_reduction','alpha':'alpha'}
+HLModelSettings['printing_items'] = {'sweeps':'sweeps','Loss':'train_loss','acc train':'train_acc',\
+												'||g||':'||g||','Loss test':'val_loss','acc test':'val_acc',\
+												'maxacc':'max_val_acc','var red':'val_variance_reduction','alpha':'alpha'}
 
 HLModel = HessianlearnModel(problem,regularization,data,settings = HLModelSettings)
